Fixed veteran experience thing I think
